% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calc_bai}
\alias{calc_bai}
\title{calculate annual basal area increment}
\usage{
calc_bai(dt.long, id, rw)
}
\arguments{
\item{dt.long}{data in long format containing at least 3 columns: id, year, rw}

\item{id}{column name of series id}

\item{rw}{column name of ring width measurement which is in mm}
}
\value{
add 3 columns to the original input data, ageC for cambial age, ba_cm2_t_1 for basal area of the previous year in cm2, and bai_cm2 for annual basal area increment in cm2
}
\description{
calculate basal area (cm2) and basal area increment (cm2)
}
