% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{ccf_avg}
\alias{ccf_avg}
\title{ccf with chronologies}
\usage{
ccf_avg(icol, data, blcrit = 0.1, lag.max = 10, qa_code = "fail")
}
\arguments{
\item{icol}{column index}

\item{data}{data in wide format, first column is "Year", second column is chronologies}

\item{blcrit}{criteria for borderline}

\item{lag.max}{maximum lag for ccf}

\item{qa_code}{quality classification code (NULL, no code)}
}
\description{
ccf with chronologies
}
