% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{ccf_pairs}
\alias{ccf_pairs}
\title{pair-wise ccf}
\usage{
ccf_pairs(ts1, ts2, max.lag = 10)
}
\arguments{
\item{ts1}{: first series}

\item{ts2}{: second series}

\item{max.lag}{: maximum lag for ccf}
}
\description{
pair-wise ccf
}
