% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{CFS_qa}
\alias{CFS_qa}
\title{tree ring data quality classification and master chronology}
\usage{
CFS_qa(
  dt.input,
  mtd.paired_ccf = "S",
  batch_size = 5000,
  max.lag = 10,
  max.trial = 100
)
}
\arguments{
\item{dt.input}{tree ring data with at least 3 columns (SampleID, Year, RawRing)}

\item{mtd.paired_ccf}{processing approach for paired ccf c("S", "LB", "LP")}

\item{batch_size}{batch size for processing large dataset}

\item{max.lag}{maximum lag for ccf}

\item{max.trial}{maximum loop for step 2 to form the master dendrochronology}
}
\value{
dt.ccf:	A data table containing the CCF results for all samples, including the quality assessment code (qa_code).

master:	The final master chronology, including both raw master chronology, the mean of ring measurement of the series with pass, and the treated  master chronology, calculated as the difference of two consecutive raw master chronology.

plot.lst:	plots of the raw and treated ring measurements with year, along with the CCF plots for each sample.
}
\description{
classify tree ring data in five classes ("pass","borderline", "pm1", "highpeak", "fail") using the difference series.
}
\details{
We present a simple algorithm to assess measurement accuracy using a treated series based on the differences between two consecutive tree-ring measurements.
The algorithm consists of two main steps:

 Step 1. Initial treated master chronology

 Perform pairwise CCF on the treated series of all possible combinations of the samples.
 The pair which meets the criteria is qualified for the master chronology calculation.
 We then calculate the raw ring master chronology by averaging the ring measurements of these selected samples.
 The difference series of the raw ring master chronology is used as the initial treated master chronology for next step.

 Step 2. Refinement on Master Chronology

 Perform CCF between the treated series of each sample and the initial treated master chronology.
 Samples that do not meet the criteria will be removed from the recalculation of the treated master chronology.

 This process is repeated until all remaining samples in the master chronology meet the criteria.

This process results in five categories classification for all samples (pass, borderline, pm1, highpeak, fail)
}
