% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{CFS_qa_old}
\alias{CFS_qa_old}
\title{tree-ring data measurement assessment}
\usage{
CFS_qa_old(
  dt.input,
  qa.label_data = "",
  qa.label_trt = "",
  qa.batch_size = 10000,
  qa.max_lag = 10,
  qa.max_iter = 100,
  qa.min_nseries = 100,
  qa.blcrit = 0.1
)
}
\arguments{
\item{dt.input}{tree ring data with at least 3 columns (SampleID, Year, RawRing)}

\item{qa.label_data}{description of dt.input}

\item{qa.label_trt}{description of treated series}

\item{qa.batch_size}{number of pairs to run in a batch, to avoid memory issues in processing large dataset}

\item{qa.max_lag}{maximum lag up to which the correlation should be calculated in CCF}

\item{qa.max_iter}{maximum number of iterations of step 2(see Details)}

\item{qa.min_nseries}{minimum number of series to run this function}

\item{qa.blcrit}{criteria for borderline}
}
\value{
A list of 3 elements:
1) dt.ccf:	A data table containing the CCF results for all samples, including the quality assessment code (qa_code).
2) dt.chron:	The final chronologies, including both raw chronologies, the mean of ring measurement of the series with pass,
and the treated  chronologies, calculated as the difference of two consecutive raw chronologies.
3) dt.stats:	summary statistics of radii
4) dt.plots, list of tables for generating plots
5) qa.parms: parameters used
}
\description{
Assess tree-ring measurement accuracy using a treated series based on the differences between two consecutive tree-ring measurements.
}
\details{
Assess tree-ring measurement accuracy using a treated series based on the differences between two consecutive tree-ring measurements.
The algorithm consists of two main steps:

 Step 1. Perform pairwise CCF on the treated series of all possible combinations of the samples.
raw ring chronologies is calculated as the average of the ring measurements of samples that achieve the maximum correlation at lag 0 with at least one other sample in the cross-correlation function (CCF).
 The difference series of the raw ring chronologies is used as the initial treated chronologies for next step.

 Step 2. Perform CCF between the treated series of each sample and the initial treated chronologies.
 Samples that do not meet the criteria will be removed from the recalculation of the treated chronologies.
 This step is repeated until all remaining samples in the chronologies meet the criteria.

This process results in five categories classification for all samples (pass, borderline, pm1, highpeak, fail)


To enhance efficiency and mitigate potential memory issues,
The function supports both parallel (multi-session) and sequential modes, and also offers a batch processing option for users.
}
