% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_uids.R
\name{count_uids}
\alias{count_uids}
\title{data spatial distribution by species measured later than certain year}
\usage{
count_uids(
  data,
  yend,
  resol_lat,
  resol_lon,
  uid,
  N.species = 999,
  make.plot = FALSE,
  out.csv = FALSE,
  out.dir = NULL
)
}
\arguments{
\item{data}{:  input 2 tables: header and rw}

\item{yend}{: year criteria}

\item{resol_lat}{:latitude resolution}

\item{resol_lon}{:longitude resolution}

\item{uid}{: counts on which uid}

\item{N.species}{: number of top species to be output (999 for all)}

\item{make.plot}{plot the distribution (TRUE or FALSE)}

\item{out.csv}{: output csv file (TRUE or FALSE)}

\item{out.dir}{: directory of output csv files}
}
\value{
1 table species-lat * lon, values are number of uids
}
\description{
data spatial distribution by species measured later than certain year
}
