% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSformat.R
\name{dt.sep2}
\alias{dt.sep2}
\title{seperate 1 table into 2: dt.meta and dt.rwl with the link id_radius.tmp}
\usage{
dt.sep2(data, N.cols_meta = NULL)
}
\arguments{
\item{data}{: dataframe or datatable in wide format}

\item{N.cols_meta}{: all column index except ring measurement}
}
\value{
2 tables: dt.meta and dt.rwl
}
\description{
seperate 1 table into 2: dt.meta and dt.rwl with the link id_radius.tmp
}
