% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{format_byterm}
\alias{format_byterm}
\title{format wide to long}
\usage{
format_byterm(model, dt.pred)
}
\arguments{
\item{model}{a gam model}
}
\value{
dt.pred, the term prediction by gam were formed as 1 term, for both fits and stand error
}
\description{
convert the format of term prediction with by from wide to long in model$pred
}
