% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gam_mod}
\alias{gam_mod}
\title{a general gam model}
\usage{
gam_mod(data, resp_scale = "", m.candidates)
}
\arguments{
\item{data}{data containing all necessary columns to run the model}

\item{resp_scale}{the scale of response variable. default is original scale with family = Gamma("log"); "log" for predicting log-scale with family = gaussian("identity") ; "hybrid" for predicting original scale with family = gaussian("identity")}

\item{m.candidates}{the list of formulas.}
}
\value{
list including model, fitting statistics, ptable, stable and prediction table
}
\description{
gam model benefiting the model selection procedure using ML method
}
\details{
This function models a gam model without considering random effects or autocorrelation.
If users specify multiple candidate models through the m.candidates argument, the function will fit each candidate model using the maximum likelihood (ML) method.
The Akaike Information Criterion (AIC) will then be compared to determine the best-fitting model. Once the optimal model is identified,
it will be refitted using the restricted maximum likelihood (REML) method and output the results.

If users specify only 1 candidate model through the m.candidates argument, the model is fitted with "REML" method.
}
