% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gamm_site}
\alias{gamm_site}
\title{growth model at site-level}
\usage{
gamm_site(data, resp_scale = "log", m.candidates)
}
\arguments{
\item{data}{data containing all necessary columns to run the model}

\item{resp_scale}{the scale of response variable. default is "log" for log-scale, otherwise modeling the response variable as it is}

\item{m.candidates}{the list of formulas.}
}
\value{
list including model, fitting statistics, ptable, stable and prediction table
}
\description{
models the growth trend or climate-growth relationship at site-level
}
\details{
This function accounts for within-site variability and temporal autocorrelation by including series identity as random effects
and a first-order autoregressive (AR1) correlation structures, respectively. “Normalized” residuals are also provided for future analysis.

If users specify multiple candidate models through the m.candidates argument, the function will fit each candidate model using the maximum likelihood (ML) method.
The Akaike Information Criterion (AIC) will then be compared to determine the best-fitting model.
Once the optimal model is identified, it will be refitted using the restricted maximum likelihood (REML) method and output the results.

If users specify only 1 candidate model through the m.candidates argument, the model is fitted with "REML" method.
}
