% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{generate_plots}
\alias{generate_plots}
\title{generating plots for all series}
\usage{
generate_plots(master.trt, dt.input)
}
\arguments{
\item{master.trt}{: master chronology from tr.fullmaster function}

\item{dt.input}{: tables from pre.dataFormat function}
}
\description{
generating plots for all series
}
