% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_freq}
\alias{plot_freq}
\title{plot frequency distribution by geo-location per species}
\usage{
plot_freq(dt.freq, out.species = "all")
}
\arguments{
\item{dt.freq}{a table resulting from function CFS_freq()}

\item{out.species}{species list, default is 'all' to output all species}
}
\description{
This function plots the frequency distribution by geo-location per species
}
