% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_qa}
\alias{plot_qa}
\title{plotting quality assessment}
\usage{
plot_qa(qa.trt, qa.out_series = "all")
}
\arguments{
\item{qa.trt}{object "cfs_qa" from CFS_qa() function}

\item{qa.out_series}{SampleID list to be plotted, default -999 for output all samples}
}
\description{
plotting ccf results for each tree
}
