% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_results}
\alias{plot_results}
\title{plotting ccf}
\usage{
plot_results(plot.lst, out_series = "A", out_N = NULL, outpdf_YN, outpdf_fn)
}
\arguments{
\item{plot.lst}{: plot list from generate_plots}

\item{out_series}{: 3 options(A: all series; P: specific series of out_N; R: randomly selected samples)}

\item{out_N}{: for out_series A, not necessary; for P: list of index of specific series; R: number of series to be randomly selected}

\item{outpdf_YN}{:  option to output the results in pdf ("Y", "N")}

\item{outpdf_fn}{: output file name including path/filename.pdf}
}
\description{
plotting ccf results for each tree
}
