% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_scale}
\alias{plot_scale}
\title{plot rw_median of the site and its neighbours on temperoral and spatial space}
\usage{
plot_scale(dt.scale)
}
\arguments{
\item{dt.scale}{a table resulting from function CFS_scale()}
}
\description{
This function plots rw_median of the site and its neighbours on temperoral and spatial space
}
