% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plots_ds}
\alias{plots_ds}
\title{plot data summary and location}
\usage{
plots_ds(data)
}
\arguments{
\item{data}{a list of 2 tables : 1st is meta data with 1 species only ; 2nd is the ring width measurement in long-format}
}
\description{
This function plots the site location and frequency distribution of series length and ring width measurement per species. It's used for generating the data report
}
