% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plots_facet}
\alias{plots_facet}
\title{plot multiple series by wrap_facet, and put them in a list}
\usage{
plots_facet(data, varcols, xylabels, nrow, ncol)
}
\arguments{
\item{data}{a data table including all the necessary columns in varcols}

\item{varcols}{a list containing 3 column names for x, y and facet category.}

\item{xylabels}{a list containing the labels for x-axis and y-axis}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per col}
}
\description{
This function plots dynamic xyplot per category using wrap_facet. each facet contains nrow*ncol plots. It's used for generating the data report
}
