% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plots_qa}
\alias{plots_qa}
\title{plotting quality assessment}
\usage{
plots_qa(plot.lst, out.series = "all", out.pdf = NULL)
}
\arguments{
\item{plot.lst}{plot list from generate_plots}

\item{out.series}{SampleID list to be plotted, default -999 for output all samples}

\item{out.pdf}{output file name including path/filename.pdf, default NULL for device screen}
}
\description{
plotting ccf results for each tree
}
