% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{pre_dataFormat}
\alias{pre_dataFormat}
\title{pre-processing data in wide and long format for rawring and treated series}
\usage{
pre_dataFormat(dt.rw_long)
}
\arguments{
\item{dt.rw_long}{: original data with at least 3 columns (SampleID, Year, RawRing)}
}
\description{
pre-processing data in wide and long format for rawring and treated series
}
