% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rwl.R
\name{read_rwl}
\alias{read_rwl}
\title{read rwl files which were downloaded locally from ITRDB}
\usage{
read_rwl(dir.src, rwl)
}
\arguments{
\item{dir.src}{directory of source rwl files}

\item{rwl}{file name}
}
\value{
two tables: header and rwl measurement
}
\description{
read rwl files which were downloaded locally from ITRDB
}
