% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_tables.R
\name{table_spc}
\alias{table_spc}
\title{Generate table for species report}
\usage{
table_spc(tr_6i, dt.spc_radii)
}
\arguments{
\item{tr_6i}{data table of meta data for a specific species}

\item{dt.spc_radii}{data table of tree ring data for a specific species}
}
\value{
a data table.
}
\description{
This function generates the data table for data reporting at project-species level .
}
