% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_tables.R
\name{table_spc_site}
\alias{table_spc_site}
\title{Generate table for site report including the kNN result}
\usage{
table_spc_site(dt.radii.spc, rw_ref, scale.N_nbs)
}
\arguments{
\item{dt.radii.spc}{result table from table_spc_site_radii}

\item{rw_ref}{data table of reference for kNN, in wide format}

\item{scale.N_nbs}{number of neighbor to search}
}
\value{
a data table.
}
\description{
This function generates the data table for data reporting at project-species-site level .
}
