% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_tables.R
\name{table_spc_site_radii}
\alias{table_spc_site_radii}
\title{Generate table for radii report}
\usage{
table_spc_site_radii(tr_6i, tr_7i, min.nseries.qa, rep_radii = TRUE)
}
\arguments{
\item{tr_6i}{data table of meta data for a specific species}

\item{tr_7i}{data table of tree ring data for a specific species}

\item{min.nseries.qa}{minimum number of series to run the qa process}

\item{rep_radii}{logic, for user to choose if the radii table is presented in the data report.}
}
\value{
a data table.
}
\description{
This function generates the data table for data reporting at project-species-site-radii level .
}
