% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{tr.fullmaster}
\alias{tr.fullmaster}
\title{tree ring data quality classification}
\usage{
tr.fullmaster(
  dt.trt_wide,
  dt.rw_long,
  mtd.paired_ccf = "S",
  batch_size = 5000,
  max.lag = 10,
  max.trial = 100
)
}
\arguments{
\item{dt.trt_wide}{: treated series in wide format for pair-wise ccf}

\item{dt.rw_long}{: ring width series in long format for calculating mean of master chronology}

\item{mtd.paired_ccf}{: processing approach for paired ccf c("S", "LB", "LP")}

\item{batch_size}{:  batch size for processing large dataset}

\item{max.lag}{: maximum lag for ccf}

\item{max.trial}{: maximum loop for step 2 to form the master dendrochronology}
}
\value{
list including model, fitting statistics, ptable, stable and prediction table
}
\description{
classify tree ring data in 4 classes ("pass","borderline", "highpeak", "pm1") using the detrended series based on 2 criteria: 1: treated
}
