% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{tr_fullmaster}
\alias{tr_fullmaster}
\title{tree ring data quality classification and master chronology}
\usage{
tr_fullmaster(
  dt.input,
  mtd.paired_ccf = "S",
  batch_size = 5000,
  max.lag = 10,
  max.trial = 100
)
}
\arguments{
\item{dt.input}{: tables from pre_dataFormat}

\item{mtd.paired_ccf}{: processing approach for paired ccf c("S", "LB", "LP")}

\item{batch_size}{:  batch size for processing large dataset}

\item{max.lag}{: maximum lag for ccf}

\item{max.trial}{: maximum loop for step 2 to form the master dendrochronology}
}
\value{
list including model, fitting statistics, ptable, stable and prediction table
}
\description{
classify tree ring data in 4 classes ("pass","borderline", "highpeak", "pm1") using the detrended series based on 2 criteria: 1: treated
}
