% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{ccf_avg}
\alias{ccf_avg}
\title{ccf with master/mean chronology}
\usage{
ccf_avg(icol, data, blcrit = 0.1, max.lag = 10, qa_code = "Fail")
}
\arguments{
\item{icol}{: column index}

\item{data}{: data in wide format, first column is "Year", second column is master/mean chronology}

\item{blcrit}{: criteria for borderline}

\item{max.lag}{: maximum lag for ccf}

\item{qa_code}{: quality classification code (NULL, no code)}
}
\description{
ccf with master/mean chronology
}
