% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSformat.R
\name{CFS_scale}
\alias{CFS_scale}
\title{compare the median of tree ring measurement of a specific site to those of nearby sites}
\usage{
CFS_scale(
  site2chk,
  ref_sites,
  max.dist_km = 20,
  N.nbs = 10,
  make.plot = c(FALSE, "")
)
}
\arguments{
\item{site2chk}{data.table with columns uid_site, site_id, species, longitude and latitude}

\item{ref_sites}{reference sites including species, uid_site, latitude, longitude, uid_radius, year, rw_mm in long format}

\item{max.dist_km}{maximum distance to search the neighbors in km}

\item{N.nbs}{number of nearest-neighbors (maximum)}

\item{make.plot}{a list , first item for plot figures or not, second item for the caption of data source}
}
\value{
A data table containing the median ring-width measurements of the involved sites, along with the distances from the specific site
}
\description{
Apply a k-nearest neighbors (k-NN) method based on geographic location for the same species.
It compares the median tree-ring measurements of a specific site to those of nearby sites
}
