% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_site.R
\name{detrend_site}
\alias{detrend_site}
\title{species-site level of detrending BAI using gamm}
\usage{
detrend_site(
  data,
  resp_scale = "log",
  mtd,
  m.candidates,
  out.mod = FALSE,
  out.csv = FALSE,
  out.dir = NULL
)
}
\arguments{
\item{data}{: datatable}

\item{resp_scale}{log or original scale of bai}

\item{mtd}{: method ML or REML}

\item{m.candidates}{: the list of RHS of formulas}

\item{out.mod}{: output gamm object (TRUE or FALSE)}

\item{out.csv}{: output csv files (TRUE or FALSE)}

\item{out.dir}{: directory of output csv files}
}
\value{
list including model, fitting statistics, ptable, stable and prediction table
}
\description{
detrending BAI using gamm model with family = Gamma(link='log') and corAR1."REML" and "ML" methods are available.It can be used to compare models with aic using "ML" method.
}
