% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{generate_report}
\alias{generate_report}
\title{Generate a data/model Report}
\usage{
generate_report(input, usage = 1, output_file = NULL)
}
\arguments{
\item{input}{a list, the first is either "data" or "model", the second is the r object, either a data table or a model object}

\item{usage}{for data report, 1 for data submission, hence the complete report; 2 for modelling use, simplified report. It has no effect on model report.}

\item{output_file}{A string specifying the name of the output HTML file.}
}
\value{
An HTML file containing the data report.
}
\description{
This function generates an HTML data report using an R Markdown template.
}
