% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{generate_report}
\alias{generate_report}
\title{Generate a data/model Report}
\usage{
generate_report(robj, usage = 1, output_file = NULL)
}
\arguments{
\item{robj}{an r object, either a list of data tables running from CFS_format()  or a list running from gamm_series, gamm_site, gamm_spatial or bam_spatial}

\item{usage}{for data report, 1 for data submission, hence the complete report; 2 for modelling use, simplified report. It has no effect on model report.}

\item{output_file}{A string specifying the name of the output HTML file.}
}
\value{
An HTML file containing the data report.
}
\description{
This function generates an HTML data report using an R Markdown template.
}
