% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_tables.R
\name{generate_report}
\alias{generate_report}
\title{Generate a data/model Report}
\usage{
generate_report(
  robj,
  reports.sel = c(1, 2, 3, 4),
  label_trt = "differentiated",
  parms.qa = c(min.nseries.qa = 20, max.dist_km = 20, N.nbs.qa = 6),
  output_file = NULL
)
}
\arguments{
\item{robj}{an r object, either a list of data tables running from CFS_format()  or a list running from gamm_series, gamm_site, gamm_spatial or bam_spatial}

\item{reports.sel}{for data report, 1 for data submission, hence the complete report; 2 for modelling use, simplified report. It has no effect on model report.}

\item{label_trt}{label of treated series for qa process}

\item{parms.qa}{a numeric vector with 2 values: minimum number of series for qa and number of neighbor for kNN}

\item{output_file}{A string specifying the name of the output HTML file.}
}
\value{
An HTML file containing the data report.
}
\description{
This function generates an HTML data report using an R Markdown template.
}
