% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFSvars.R
\name{variables.cfstrend}
\alias{variables.cfstrend}
\title{prepare variable names and column order for each tables}
\usage{
variables.cfstrend()
}
\value{
two lists: meta.all and ord.all
}
\description{
prepare variable names and column order for each tables
}
\examples{
meta.all <- variables.cfstrend()
}
